+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Next
 
-[diff](https://github.com/prettier/plugin-pug/compare/1.1.7...master)
+[diff](https://github.com/prettier/plugin-pug/compare/1.1.8...master)
+
+# 1.1.8
+
+[diff](https://github.com/prettier/plugin-pug/compare/1.1.7...1.1.8)
+
+- Calculate correct line width ([26fefff])
+- Fix indentation for `doctype` ([51d3860])
+
+[26fefff]: https://github.com/prettier/plugin-pug/commit/26fefffbe85f3f698545dbdd68c16140f4afd38c
+[51d3860]: https://github.com/prettier/plugin-pug/commit/51d386083eb1a24f57127267d6b486416b94b1c4
 
 # 1.1.7